Made Program work without ConsoleProgram
The program now uses the BufferedReader class to read User Input in the IDE's built in console and no longer uses ConsoleProgram as a super class. Input mismatches are caught using try-catch. Should the user enter a wrong input type, eg. "zweitausend" for 'year', a error message is displayed and the method is restarted, so that the user can provide a new input.